Initial Commit
This commit includes the source code for the implementation of the NN model and the weights that achieved 72.87% accuracy on the testing set. The commit also includes CCRN(Accuracy per class) testing code. The weight file is compressed into 7 .7z files to get the .h5 file extract the 001 file in the working directory. The h5 file will be used as a pretrained model.